Found and fixed daytypes bug
daytypes column was not being transfered over correctly. It was being
insert into the dataframe as a timeseries, but it had different index
values so it was not being inserted anywhere. Could have been fixed by
converting it to a list, but this way is safer w.r.t. different number
of elements that might come up with bad data later.